Move Noris functionality from `admin.service.ts` to Noris module and refactor dependencies.
