+++ nest-tax-backend/src/admin/admin.service.ts
@@ -1,11 +1,7 @@
 import { Injectable } from '@nestjs/common'
-import { PaymentStatus, Prisma, Tax, TaxPayment } from '@prisma/client'
-import currency from 'currency.js'
-import { ResponseUserByBirthNumberDto } from 'openapi-clients/city-account'
 
 import { BloomreachService } from '../bloomreach/bloomreach.service'
-import { NorisPaymentsDto, NorisTaxPayersDto } from '../noris/noris.dto'
-import { CustomErrorNorisTypesEnum } from '../noris/noris.errors'
+import { NorisPaymentsDto } from '../noris/noris.dto'
 import { NorisService } from '../noris/noris.service'
 import {
   DeliveryMethod,
@@ -14,15 +10,10 @@ import {
 } from '../noris/utils/noris.types'
 import { PrismaService } from '../prisma/prisma.service'
 import { addSlashToBirthNumber } from '../utils/functions/birthNumber'
-import { ErrorsEnum, ErrorsResponseEnum } from '../utils/guards/dtos/error.dto'
+import { ErrorsEnum } from '../utils/guards/dtos/error.dto'
 import ThrowerErrorGuard from '../utils/guards/errors.guard'
 import { CityAccountSubservice } from '../utils/subservices/cityaccount.subservice'
 import { LineLoggerSubservice } from '../utils/subservices/line-logger.subservice'
-import { QrCodeSubservice } from '../utils/subservices/qrcode.subservice'
-import {
-  TaxIdVariableSymbolYear,
-  TaxWithTaxPayer,
-} from '../utils/types/types.prisma'
 import {
   NorisRequestGeneral,
   RequestAdminCreateTestingTaxDto,
@@ -31,14 +22,6 @@ import {
   RequestUpdateNorisDeliveryMethodsDto,
 } from './dtos/requests.dto'
 import { CreateBirthNumbersResponseDto } from './dtos/responses.dto'
-import {
-  convertCurrencyToInt,
-  mapNorisToTaxAdministratorData,
-  mapNorisToTaxData,
-  mapNorisToTaxDetailData,
-  mapNorisToTaxInstallmentsData,
-  mapNorisToTaxPayerData,
-} from '../noris/utils/mapping.helper'
 import { createTestingTaxMock } from './utils/testing-tax-mock'
 
 @Injectable()
@@ -47,7 +30,6 @@ export class AdminService {
 
   constructor(
     private readonly prismaService: PrismaService,
-    private readonly qrCodeSubservice: QrCodeSubservice,
     private readonly cityAccountSubservice: CityAccountSubservice,
     private readonly bloomreachService: BloomreachService,
     private readonly norisService: NorisService,
@@ -56,353 +38,18 @@ export class AdminService {
     this.logger = new LineLoggerSubservice(AdminService.name)
   }
 
-  private async insertTaxPayerDataToDatabase(
-    dataFromNoris: NorisTaxPayersDto,
-    year: number,
-    transaction: Prisma.TransactionClient,
-    userDataFromCityAccount: ResponseUserByBirthNumberDto | null,
-  ) {
-    const taxAdministratorData = mapNorisToTaxAdministratorData(dataFromNoris)
-    const taxAdministrator = await transaction.taxAdministrator.upsert({
-      where: {
-        id: dataFromNoris.vyb_id,
-      },
-      create: taxAdministratorData,
-      update: taxAdministratorData,
-    })
-
-    const taxPayerData = mapNorisToTaxPayerData(dataFromNoris, taxAdministrator)
-    const taxPayer = await transaction.taxPayer.upsert({
-      where: {
-        birthNumber: dataFromNoris.ICO_RC,
-      },
-      create: taxPayerData,
-      update: taxPayerData,
-    })
-
-    const [qrCodeEmail, qrCodeWeb] = await Promise.all([
-      this.qrCodeSubservice.createQrCode({
-        amount: convertCurrencyToInt(dataFromNoris.dan_spolu),
-        variableSymbol: dataFromNoris.variabilny_symbol,
-        specificSymbol: '2024100000',
-      }),
-      this.qrCodeSubservice.createQrCode({
-        amount: convertCurrencyToInt(dataFromNoris.dan_spolu),
-        variableSymbol: dataFromNoris.variabilny_symbol,
-        specificSymbol: '2024200000',
-      }),
-    ])
-
-    // deliveryMethod is missing here, since we do not want to update
-    // historical taxes with the current delivery method in Noris
-    const taxData = mapNorisToTaxData(
-      dataFromNoris,
-      year,
-      taxPayer.id,
-      qrCodeEmail,
-      qrCodeWeb,
-    )
-    const tax = await transaction.tax.upsert({
-      where: {
-        taxPayerId_year: {
-          taxPayerId: taxPayer.id,
-          year,
-        },
-      },
-      update: taxData,
-      create: {
-        ...taxData,
-        deliveryMethod: userDataFromCityAccount?.taxDeliveryMethodAtLockDate,
-      },
-    })
-
-    const taxInstallments = mapNorisToTaxInstallmentsData(dataFromNoris, tax.id)
-    await transaction.taxInstallment.createMany({
-      data: taxInstallments,
-    })
-
-    const taxDetailData = mapNorisToTaxDetailData(dataFromNoris, tax.id)
-
-    await transaction.taxDetail.createMany({
-      data: taxDetailData,
-    })
-    return taxPayer
-  }
-
-  async processNorisTaxData(
-    norisData: NorisTaxPayersDto[],
-    year: number,
-  ): Promise<string[]> {
-    const birthNumbersResult: Set<string> = new Set()
-
-    this.logger.log(`Data loaded from noris - count ${norisData.length}`)
-
-    const userDataFromCityAccount =
-      await this.cityAccountSubservice.getUserDataAdminBatch(
-        norisData.map((norisRecord) => norisRecord.ICO_RC),
-      )
-
-    const taxesExist = await this.prismaService.tax.findMany({
-      select: {
-        taxPayer: {
-          select: {
-            birthNumber: true,
-          },
-        },
-      },
-      where: {
-        year: +year,
-        taxPayer: {
-          birthNumber: {
-            in: norisData.map((norisRecord) => norisRecord.ICO_RC),
-          },
-        },
-      },
-    })
-    const birthNumbersWithExistingTax = new Set(
-      taxesExist.map((tax) => tax.taxPayer.birthNumber),
-    )
-
-    await Promise.all(
-      norisData.map(async (norisItem) => {
-        birthNumbersResult.add(norisItem.ICO_RC)
-
-        const taxExists = birthNumbersWithExistingTax.has(norisItem.ICO_RC)
-        if (taxExists) {
-          return
-        }
-
-        try {
-          await this.prismaService.$transaction(async (tx) => {
-            const userFromCityAccount =
-              userDataFromCityAccount[norisItem.ICO_RC] || null
-
-            const userData = await this.insertTaxPayerDataToDatabase(
-              norisItem,
-              year,
-              tx,
-              userFromCityAccount,
-            )
-
-            if (userFromCityAccount) {
-              const bloomreachTracker =
-                await this.bloomreachService.trackEventTax(
-                  {
-                    amount: convertCurrencyToInt(norisItem.dan_spolu),
-                    year,
-                    delivery_method:
-                      userFromCityAccount.taxDeliveryMethodAtLockDate ?? null,
-                  },
-                  userFromCityAccount.externalId ?? undefined,
-                )
-              if (!bloomreachTracker) {
-                throw this.throwerErrorGuard.InternalServerErrorException(
-                  ErrorsEnum.INTERNAL_SERVER_ERROR,
-                  `Error in send Tax data to Bloomreach for tax payer with ID ${userData.id} and year ${year}`,
-                )
-              }
-            }
-          })
-        } catch (error) {
-          this.logger.error(
-            this.throwerErrorGuard.InternalServerErrorException(
-              ErrorsEnum.INTERNAL_SERVER_ERROR,
-              'Failed to insert tax to database.',
-              undefined,
-              undefined,
-              error,
-            ),
-          )
-
-          // Remove the birth number from the result set if insertion fails
-          birthNumbersResult.delete(norisItem.ICO_RC)
-        }
-      }),
-    )
-
-    // Add the payments for these added taxes to database
-    await this.updatePaymentsFromNorisWithData(norisData)
-
-    return [...birthNumbersResult]
-  }
-
   async loadDataFromNoris(
     data: RequestPostNorisLoadDataDto,
   ): Promise<CreateBirthNumbersResponseDto> {
-    this.logger.log('Start Loading data from noris')
-    const norisData = (await this.norisService.getTaxDataByYearAndBirthNumber(
+    return this.norisService.getAndProcessNewNorisTaxDataByBirthNumberAndYear(
       data,
-    )) as NorisTaxPayersDto[]
-
-    const birthNumbersResult: string[] = await this.processNorisTaxData(
-      norisData,
-      data.year,
     )
-
-    return { birthNumbers: birthNumbersResult }
   }
 
   async updateDataFromNoris(data: RequestPostNorisLoadDataDto) {
-    let norisData: NorisTaxPayersDto[]
-    try {
-      norisData = (await this.norisService.getTaxDataByYearAndBirthNumber(
-        data,
-      )) as NorisTaxPayersDto[]
-    } catch (error) {
-      throw this.throwerErrorGuard.InternalServerErrorException(
-        CustomErrorNorisTypesEnum.GET_TAXES_FROM_NORIS_ERROR,
-        'Failed to get taxes from Noris',
-        undefined,
-        undefined,
-        error,
-      )
-    }
-    let count = 0
-
-    const userDataFromCityAccount =
-      await this.cityAccountSubservice.getUserDataAdminBatch(
-        norisData.map((norisRecord) => norisRecord.ICO_RC),
-      )
-
-    const taxesExist = await this.prismaService.tax.findMany({
-      select: {
-        id: true,
-        taxPayer: {
-          select: {
-            birthNumber: true,
-          },
-        },
-      },
-      where: {
-        year: data.year,
-        taxPayer: {
-          birthNumber: {
-            in: norisData.map((norisRecord) => norisRecord.ICO_RC),
-          },
-        },
-      },
-    })
-    const birthNumberToTax = new Map(
-      taxesExist.map((tax) => [tax.taxPayer.birthNumber, tax]),
+    return await this.norisService.getNorisTaxDataByBirthNumberAndYearAndUpdateExistingRecords(
+      data,
     )
-
-    await Promise.all(
-      norisData.map(async (norisItem) => {
-        const taxExists = birthNumberToTax.get(norisItem.ICO_RC)
-        if (taxExists) {
-          try {
-            await this.prismaService.$transaction(async (tx) => {
-              await tx.taxInstallment.deleteMany({
-                where: {
-                  taxId: taxExists.id,
-                },
-              })
-              await tx.taxDetail.deleteMany({
-                where: {
-                  taxId: taxExists.id,
-                },
-              })
-
-              const userFromCityAccount =
-                userDataFromCityAccount[norisItem.ICO_RC] || null
-
-              const userData = await this.insertTaxPayerDataToDatabase(
-                norisItem,
-                data.year,
-                tx,
-                userFromCityAccount,
-              )
-              if (userData) {
-                count += 1
-              }
-            })
-          } catch (error) {
-            this.logger.error(
-              this.throwerErrorGuard.InternalServerErrorException(
-                ErrorsEnum.INTERNAL_SERVER_ERROR,
-                'Failed to update tax in database.',
-                undefined,
-                undefined,
-                error,
-              ),
-            )
-          }
-        }
-      }),
-    )
-
-    return { updated: count }
-  }
-
-  private async createTaxMapByVariableSymbol(
-    norisPaymentData: Partial<NorisPaymentsDto>[],
-  ) {
-    const taxesData = await this.prismaService.tax.findMany({
-      where: {
-        variableSymbol: {
-          in: norisPaymentData
-            .filter((item) => item.variabilny_symbol !== undefined)
-            .map((item) => item.variabilny_symbol as string),
-        },
-      },
-      include: {
-        taxPayer: true,
-      },
-    })
-    return new Map(taxesData.map((item) => [item.variableSymbol, item]))
-  }
-
-  // TODO: Eventually we want to get rid of this function, and do some better error handling, than watching these specific cases.
-  /**
-   * This function handles errors in the payment process. It logs an error message if the payment process is not correct, with the info about why it is not correct.
-   *
-   * @param paidFromNoris Already paid amount in Noris.
-   * @param taxData Tax object, containing all the information about the tax.
-   * @param forPayment Left to be paid amount in Noris.
-   * @param payerDataCountAll How many payments for this tax we have in the database.
-   */
-  private handlePaymentsErrors(
-    paidFromNoris: number,
-    taxData: Tax,
-    forPayment: number,
-    payerDataCountAll: number,
-  ) {
-    if (paidFromNoris > taxData.amount && forPayment === 0) {
-      this.logger.error(
-        this.throwerErrorGuard.InternalServerErrorException(
-          ErrorsEnum.INTERNAL_SERVER_ERROR,
-          'ZAPLATENE VSETKO ALE V NORISE JE VACSIA CIASTKA AKO U NAS',
-        ),
-      )
-    } else if (
-      payerDataCountAll === 0 &&
-      paidFromNoris >= taxData.amount &&
-      forPayment > 0
-    ) {
-      this.logger.error(
-        this.throwerErrorGuard.InternalServerErrorException(
-          ErrorsEnum.INTERNAL_SERVER_ERROR,
-          'U NAS ZAPLATENE VSETKO ALE V NORISE NIE - na 1x',
-        ),
-      )
-    } else if (
-      payerDataCountAll > 0 &&
-      paidFromNoris >= taxData.amount &&
-      forPayment > 0
-    ) {
-      this.logger.error(
-        this.throwerErrorGuard.InternalServerErrorException(
-          ErrorsEnum.INTERNAL_SERVER_ERROR,
-          'U NAS ZAPLATENE VSETKO ALE V NORISE NIE - na x krat',
-        ),
-      )
-    }
-  }
-
-  private formatAmount(amount: number | string) {
-    return typeof amount === 'number'
-      ? currency(amount).intValue
-      : convertCurrencyToInt(amount)
   }
 
   async updatePaymentsFromNoris(norisRequest: NorisRequestGeneral) {
@@ -412,190 +59,7 @@ export class AdminService {
         : await this.norisService.getPaymentDataFromNorisByVariableSymbols(
             norisRequest.data,
           )
-    return this.updatePaymentsFromNorisWithData(norisPaymentData)
-  }
-
-  private async processNorisPaymentData(
-    norisPaymentData: Partial<NorisPaymentsDto>[],
-    taxesDataByVsMap: Map<string, TaxWithTaxPayer>,
-    taxPaymentDataMap: Map<number, { sum: number; count: number }>,
-    userDataFromCityAccount: Record<string, ResponseUserByBirthNumberDto> = {},
-  ) {
-    const validPayments = norisPaymentData.filter(
-      (norisPayment) =>
-        norisPayment.variabilny_symbol !== undefined &&
-        norisPayment.uhrazeno !== undefined &&
-        norisPayment.zbyva_uhradit !== undefined,
-    )
-
-    // Step 2: Process each payment separately
-    const paymentProcesses = validPayments.map((norisPayment) =>
-      this.processIndividualPayment(
-        norisPayment,
-        taxesDataByVsMap,
-        taxPaymentDataMap,
-        userDataFromCityAccount,
-      ),
-    )
-
-    // Step 3: Execute all payment processes concurrently
-    return Promise.all(paymentProcesses)
-  }
-
-  private async processIndividualPayment(
-    norisPayment: Partial<NorisPaymentsDto>,
-    taxesDataByVsMap: Map<string, TaxWithTaxPayer>,
-    taxPaymentDataMap: Map<number, { sum: number; count: number }>,
-    userDataFromCityAccount: Record<string, ResponseUserByBirthNumberDto> = {},
-  ) {
-    try {
-      const taxData = taxesDataByVsMap.get(norisPayment.variabilny_symbol!)
-
-      if (!taxData) {
-        return 'NOT_EXIST'
-      }
-
-      const payerData = taxPaymentDataMap.get(taxData.id) || {
-        sum: 0,
-        count: 0,
-      }
-      const paidFromNoris = this.formatAmount(norisPayment.uhrazeno!)
-      const toPayFromNoris = this.formatAmount(norisPayment.zbyva_uhradit!)
-
-      // Early return if payment already recorded
-      if (payerData.sum !== null && payerData.sum >= paidFromNoris) {
-        return 'ALREADY_CREATED'
-      }
-
-      await this.prismaService.$transaction(async (tx) => {
-        const createdTaxPayment = await tx.taxPayment.create({
-          data: {
-            amount: paidFromNoris - (payerData.sum ?? 0),
-            source: 'BANK_ACCOUNT',
-            specificSymbol: norisPayment.specificky_symbol,
-            taxId: taxData.id,
-            status: PaymentStatus.SUCCESS,
-          },
-        })
-
-        await this.trackPaymentIfNeeded(
-          taxData,
-          createdTaxPayment,
-          userDataFromCityAccount,
-        )
-      })
-
-      this.handlePaymentsErrors(
-        paidFromNoris,
-        taxData,
-        toPayFromNoris,
-        payerData.count,
-      )
-      return 'CREATED'
-    } catch (error) {
-      return this.throwerErrorGuard.InternalServerErrorException(
-        ErrorsEnum.INTERNAL_SERVER_ERROR,
-        ErrorsResponseEnum.INTERNAL_SERVER_ERROR,
-        undefined,
-        undefined,
-        error,
-      )
-    }
-  }
-
-  private async trackPaymentIfNeeded(
-    taxData: TaxWithTaxPayer,
-    createdTaxPayment: TaxPayment,
-    userDataFromCityAccount: Record<string, ResponseUserByBirthNumberDto>,
-  ) {
-    const userFromCityAccount =
-      userDataFromCityAccount[taxData.taxPayer.birthNumber] || null
-
-    if (userFromCityAccount && userFromCityAccount.externalId) {
-      await this.bloomreachService.trackEventTaxPayment(
-        {
-          amount: createdTaxPayment.amount,
-          payment_source: 'BANK_ACCOUNT',
-          year: taxData.year,
-        },
-        userFromCityAccount.externalId,
-      )
-    }
-  }
-
-  async updatePaymentsFromNorisWithData(
-    norisPaymentData: Partial<NorisPaymentsDto>[],
-  ) {
-    const taxesDataByVsMap =
-      await this.createTaxMapByVariableSymbol(norisPaymentData)
-
-    // Get all tax IDs from taxesDataByVsMap
-    const taxIds = Array.from(taxesDataByVsMap.values(), (tax) => tax.id)
-
-    // Get aggregate payment data for all taxes at once
-    const taxPaymentDataMap = await this.fetchTaxPaymentAggregateMap(taxIds)
-
-    // Get batch data from city account
-    const userDataFromCityAccount =
-      await this.cityAccountSubservice.getUserDataAdminBatch(
-        Array.from(
-          taxesDataByVsMap.values(),
-          (taxData) => taxData.taxPayer.birthNumber,
-        ),
-      )
-
-    const resultList = await this.processNorisPaymentData(
-      norisPaymentData,
-      taxesDataByVsMap,
-      taxPaymentDataMap,
-      userDataFromCityAccount,
-    )
-    const created = resultList.filter((item) => item === 'CREATED').length
-    const alreadyCreated = resultList.filter(
-      (item) => item === 'ALREADY_CREATED',
-    ).length
-    const errors: Error[] = resultList.filter((item) => item instanceof Error)
-
-    if (errors.length > 0) {
-      this.logger.error(
-        'Encountered errors while batch processing Noris payments:',
-        errors,
-      )
-    }
-
-    return {
-      created,
-      alreadyCreated,
-    }
-  }
-
-  private async fetchTaxPaymentAggregateMap(taxIds: number[]) {
-    const aggregateData = await this.prismaService.taxPayment.groupBy({
-      by: ['taxId'],
-      _sum: {
-        amount: true,
-      },
-      _count: {
-        _all: true,
-      },
-      where: {
-        taxId: {
-          in: taxIds,
-        },
-        status: PaymentStatus.SUCCESS,
-      },
-    })
-
-    // Create a map of aggregated data for easy lookup
-    return new Map(
-      aggregateData.map((data) => [
-        data.taxId,
-        {
-          sum: data._sum.amount || 0,
-          count: data._count._all,
-        },
-      ]),
-    )
+    return this.norisService.updatePaymentsFromNorisWithData(norisPaymentData)
   }
 
   async updateDeliveryMethodsInNoris({
@@ -678,36 +142,6 @@ export class AdminService {
     ])
   }
 
-  async updateTaxesFromNoris(taxes: TaxIdVariableSymbolYear[]): Promise<void> {
-    const variableSymbolToId = new Map(
-      taxes.map((tax) => [tax.variableSymbol, tax.id]),
-    )
-    const variableSymbols = [...variableSymbolToId.keys()]
-    const years = [...new Set(taxes.map((tax) => tax.year))]
-    const data = await this.norisService.getDataForUpdate(
-      variableSymbols,
-      years,
-    )
-    const variableSymbolsToNonNullDateFromNoris: Map<string, string> = new Map(
-      data
-        .filter((item) => item.datum_platnosti !== null)
-        .map((item) => [
-          item.variabilny_symbol,
-          item.datum_platnosti as string,
-        ]),
-    )
-
-    await this.prismaService.$transaction(
-      [...variableSymbolsToNonNullDateFromNoris.entries()].map(
-        ([variableSymbol, dateTaxRuling]) =>
-          this.prismaService.tax.update({
-            where: { id: variableSymbolToId.get(variableSymbol) },
-            data: { dateTaxRuling },
-          }),
-      ),
-    )
-  }
-
   /**
    * Creates a testing tax record with specified details for development and testing purposes.
    * WARNING! This tax should be removed after testing, with the endpoint `delete-testing-tax`.
@@ -733,7 +167,7 @@ export class AdminService {
     )
 
     // Process the mock data to create the testing tax
-    await this.processNorisTaxData([mockTaxRecord], year)
+    await this.norisService.processNorisTaxData([mockTaxRecord], year)
   }
 
   async deleteTax({
