+++ nest-tax-backend/src/noris/noris.module.ts
@@ -7,9 +7,12 @@ import { NorisConnectionSubservice } from './subservices/noris-connection.subser
 import { NorisDeliveryMethodSubservice } from './subservices/noris-delivery-method.subservice'
 import { NorisPaymentSubservice } from './subservices/noris-payment.subservice'
 import { NorisTaxSubservice } from './subservices/noris-tax.subservice'
+import { CityAccountSubservice } from '../utils/subservices/cityaccount.subservice'
+import { BloomreachModule } from '../bloomreach/bloomreach.module'
+import { QrCodeSubservice } from '../utils/subservices/qrcode.subservice'
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, BloomreachModule],
   providers: [
     NorisService,
     NorisConnectionSubservice,
@@ -17,6 +20,8 @@ import { NorisTaxSubservice } from './subservices/noris-tax.subservice'
     NorisPaymentSubservice,
     NorisTaxSubservice,
     ThrowerErrorGuard,
+    CityAccountSubservice,
+    QrCodeSubservice,
   ],
   exports: [NorisService],
   controllers: [],
